+++ CHANGELOG.md
@@ -8,7 +8,8 @@
 - Added `MaxFunctionEvals` and `MaxGenerations` termination criteria.
 - Adjusted the signatures of `CMAES::run` and `CMAES::print_final_info` to reflect the new termination criteria.
 - Renamed `reason` field of `TerminationData` to `reasons` and changed its type from `TerminationReason` to `Vec<TerminationReason>`.
-- Change signature of `CMAES::print_final_info` to take `&[TerminationReason]` instead of `TerminationReason`.
+- Changed signature of `CMAES::print_final_info` to take `&[TerminationReason]` instead of `TerminationReason`.
+- Fixed `TolFun` termination criterion checking whether the current generation's function values are all below `tol_fun` instead of whether the range of them is.
 
 ## 0.1.1 (March 4th, 2022)